Added new file, built SQLite functionality
Split the program in two, a grabber and a tweeter. Here's the grabber.
The grabber *grabs* 3 posts from the sub, around 1/6 of them is from rising.
Upon grabbing the posts, they're stored in a SQLite db with an enforced
size limit of about 5 untweeted posts. The tweeted posts stay piled up in the db
in order to allow future posts to be checked as to whether they're a copy.
Also added a couple lines to the image downloader to give it a bit more
flexibility with unruly imgur links. Also ran it a bunch of times so hopefully
nothing explodes upon deployment.

The idea is that this script will be rigged to run once a day, while
the tweeter will be ran maybe 2-3 times a day with a bit of irregularity
thrown in. This one grabs 'em, the other one tweets 'em.
